requirements.txt
discord.py
pymongo
dnspython
requests